+++ src/updating.ts
@@ -4,10 +4,11 @@ import { copyItemIdsByItemType, copyStates } from "./copyStates";
 import { getStatesDiff } from "./getStatesDiff";
 import { updateRepondNextFrame } from "./helpers/frames";
 import { repondMeta as meta, RecordedChanges, RepondMetaPhase } from "./meta";
-import { Effect, EffectPhase } from "./types";
+import { EffectDef, EffectPhase } from "./types";
 
 const MAX_STEP_ITERATIONS = 8;
 
+// NOTE maybe update to only check propIds and also itemIds by propId or something?
 function updateDiffInfo(recordedChanges: RecordedChanges) {
   //  make a diff of the changes
   getStatesDiff(meta.nowState, meta.prevState, meta.diffInfo, recordedChanges, false /* checkAllChanges */);
@@ -72,7 +73,7 @@ function runAddAndRemove() {
   meta.addAndRemoveItemsQueue.length = 0;
 }
 
-function runEffectChangesPerItemForItemType(effect: Effect, type: string) {
+function runEffectChangesPerItemForItemType(effect: EffectDef, type: string) {
   const diffInfo = meta.diffInfo;
   const allowedIdsMap = effect._allowedIdsMap;
   const props = effect._propsByItemType?.[type];
@@ -122,6 +123,7 @@ function runEffectChangesPerItemForItemType(effect: Effect, type: string) {
 }
 
 function runEffects(phase: EffectPhase, stepName: string) {
+  // NOTE Check effects doesnt use recorded changes! onyl the diff info?
   const effectNamesToRun = checkEffects(phase, stepName);
 
   for (let index = 0; index < effectNamesToRun.length; index++) {
@@ -186,6 +188,7 @@ export function createRecordedChanges(recordedChanges: RecordedChanges) {
   recordedChanges.somethingChanged = false;
 }
 
+// Maybe this can be removed? and just the props chenge thing, and also maybe ids changed?
 function resetRecordedChanges(recordedChanges: RecordedChanges) {
   recordedChanges.somethingChanged = false;
 
@@ -219,9 +222,14 @@ function resetRecordedStepChanges() {
 // BUT I think the first time it runs, it doesn't have any diff info, so it runs the effects first which maybe won't do anything, then the setStates
 // the setStates from clalbacks only get run here and not before
 function runStepEffects(stepName: string) {
+  // NOTE this runs based on the diff info of the PREVIOUS steps END changes, which is all colected changes so far
+
+  //
+
   resetRecordedStepChanges(); // NOTE recently added to prevent 'derive' changes being remembered each time it derives again
   runEffectsWithRunAtStart(); // run the runAtStart listeners
   runEffects("duringStep", stepName); //  a running derive-listener can add more to the setStates que (or others)
+  meta.recordedPropIdsChangedMap.duringStep = {};
   runAddEffects(); // add rules / effects
   runAddAndRemove(); // add and remove items
   runSetStates(); // run the qued setStates
@@ -244,8 +252,11 @@ function removeRemovedItemRefs() {
 
 function runSetOfStepEffects(stepName: string) {
   meta.nowMetaPhase = "runningEffects";
+  meta.nowEffectPhase = "duringStep";
+  meta.isFirstDuringPhaseLoop = true;
   for (let i = 0; i < MAX_STEP_ITERATIONS; i++) {
     runStepEffects(stepName);
+    meta.isFirstDuringPhaseLoop = false;
     if (!meta.recordedEffectChanges.somethingChanged) return;
   }
   logTooManySetStatesMessage();
@@ -253,6 +264,7 @@ function runSetOfStepEffects(stepName: string) {
 
 function runStepEndEffects(stepName: string) {
   meta.nowMetaPhase = "runningStepEndEffects"; // hm not checked anywhere, but checking metaPhase !== "runningEffects" (runnin derrivers) is
+  meta.nowEffectPhase = "endOfStep";
   updateDiffInfo(meta.recordedStepEndEffectChanges); // the diff for all the combined derriver changes
   runEffects("endOfStep", stepName); //  Then it runs the stepEnd effects based on the diff
 }
@@ -293,6 +305,7 @@ export function _updateRepond(animationFrameTime: number) {
 
   runSetOfStepsLoop();
   resetRecordedStepEndChanges(); // maybe resetting recorded changes here is better, before the callbacks run? maybe it doesnt matter?
+  meta.recordedPropIdsChangedMap.endOfStep = {};
 
   setMetaPhase("waitingForFirstUpdate");
   runNextTickCallbacks();
@@ -314,7 +327,7 @@ function logTooManySetStatesMessage() {
   console.warn("WARNING: running step effects a lot, there may be an infinite setState inside an effect");
   console.log("Step name: ", meta.nowStepName);
   console.log("Effect ids:");
-  console.log(JSON.stringify(meta.effectIdsByPhaseByStep.duringStep?.[meta.nowStepName], null, 2));
+  console.log(JSON.stringify(meta.effectIdsByPhaseByStepByPropId.duringStep?.[meta.nowStepName], null, 2));
   console.log("Changes");
   console.log(JSON.stringify(getDebugStepEffectsData(), null, 2));
 }