+++ src/usable/getSet.ts
@@ -1,5 +1,4 @@
 import { forEach } from "chootils/dist/loops";
-import { mergeToState } from "../copyStates";
 import { runWhenAddingAndRemovingItems, whenSettingStates } from "../helpers/runWhens";
 import { repondMeta as meta } from "../meta";
 import {
@@ -30,6 +29,13 @@ export function setState(propPath: string, newValue: any, itemId?: string) {
       return;
     }
 
+    meta.recordedPropIdsChangedMap[meta.nowEffectPhase][propPath] = true;
+    meta.recordedPropIdsChangedMap.endOfStep[propPath] = true;
+    // if (meta.nowEffectPhase === "duringStep") {
+    // console.log("meta.nowEffectPhase", meta.nowEffectPhase);
+    // console.log(meta.recordedPropIdsChangedMap.duringStep?.[propPath]);
+    // }
+
     const storeType = meta.itemTypeByPropPathId[propPath];
     const propKey = meta.propKeyByPropPathId[propPath];
     let foundItemId = itemId || meta.itemIdsByItemType[storeType]?.[0];
@@ -40,15 +46,26 @@ export function setState(propPath: string, newValue: any, itemId?: string) {
       );
     }
 
-    mergeToState(
-      storeType,
-      propKey,
-      newValue,
-      foundItemId,
-      meta.nowState,
-      meta.nowMetaPhase === "runningEffects" ? meta.recordedEffectChanges : meta.recordedStepEndEffectChanges,
-      meta.recordedStepEndEffectChanges
-    );
+    const recordedChanges =
+      meta.nowMetaPhase === "runningEffects" ? meta.recordedEffectChanges : meta.recordedStepEndEffectChanges;
+    const allRecordedChanges = meta.recordedStepEndEffectChanges;
+
+    // check if the item exists before copying
+    if (meta.nowState?.[storeType]?.[foundItemId] === undefined) return;
+
+    // save the new state
+    meta.nowState[storeType][foundItemId][propKey] = newValue;
+
+    recordedChanges.itemTypesBool[storeType] = true;
+    recordedChanges.itemIdsBool[storeType][foundItemId] = true;
+    recordedChanges.itemPropsBool[storeType][foundItemId][propKey] = true;
+
+    allRecordedChanges.itemTypesBool[storeType] = true;
+    allRecordedChanges.itemIdsBool[storeType][foundItemId] = true;
+    allRecordedChanges.itemPropsBool[storeType][foundItemId][propKey] = true;
+
+    recordedChanges.somethingChanged = true;
+    allRecordedChanges.somethingChanged = true;
   });
 }
 
@@ -154,6 +171,8 @@ export function addItem<T_ItemType extends ItemType>(
   if (!meta.willAddItemsInfo[type]) meta.willAddItemsInfo[type] = {};
   meta.willAddItemsInfo[type][id] = true;
 
+  const propPath = `${type}.__added`;
+
   runWhenAddingAndRemovingItems(() => {
     meta.nowState[type][id] = {
       ...meta.defaultStateByItemType[type](id),
@@ -181,6 +200,9 @@ export function addItem<T_ItemType extends ItemType>(
     meta.recordedEffectChanges.itemIdsBool[type][id] = true;
     meta.recordedEffectChanges.somethingChanged = true;
 
+    meta.recordedPropIdsChangedMap[meta.nowEffectPhase][propPath] = true;
+    meta.recordedPropIdsChangedMap.endOfStep[propPath] = true;
+
     // NOTE new items with props different to the defaults props are recorded as changed
     const itemPropNames = meta.propNamesByItemType[type];
     forEach(itemPropNames, (propName) => {
@@ -197,6 +219,8 @@ export function addItem<T_ItemType extends ItemType>(
 export function removeItem(type: ItemType, id: string) {
   if (!meta.willRemoveItemsInfo[type]) meta.willRemoveItemsInfo[type] = {};
   meta.willRemoveItemsInfo[type][id] = true;
+  const propPath = `${type}.__added`;
+
   runWhenAddingAndRemovingItems(() => {
     // removing itemId
     delete meta.nowState[type][id];
@@ -206,6 +230,9 @@ export function removeItem(type: ItemType, id: string) {
     meta.recordedStepEndEffectChanges.somethingChanged = true;
     meta.recordedEffectChanges.itemTypesBool[type] = true;
     meta.recordedEffectChanges.somethingChanged = true;
+
+    meta.recordedPropIdsChangedMap[meta.nowEffectPhase][propPath] = true;
+    meta.recordedPropIdsChangedMap.endOfStep[propPath] = true;
   });
 }