+++ src/usable/hooks.ts
@@ -1,13 +1,13 @@
 import { useCallback, useEffect, useLayoutEffect, useRef, useState } from "react";
 import { toSafeEffectId } from "../helpers/effects";
 import { repondMeta as meta } from "../meta";
-import { AllState, Effect, ItemPropsByType, ItemType } from "../types";
+import { AllState, EffectDef, ItemPropsByType, ItemType } from "../types";
 import { startNewEffect, stopEffect } from "./effects";
 import { getState } from "./getSet";
 
 export function useStore<K_Type extends ItemType, T_ReturnedRepondProps>(
   whatToReturn: (diffInfo: typeof meta.diffInfo) => T_ReturnedRepondProps,
-  options: Omit<Effect, "run">,
+  options: Omit<EffectDef, "run">,
   hookDeps: any[] = []
 ): T_ReturnedRepondProps {
   const [, setTick] = useState(0);
@@ -22,7 +22,7 @@ export function useStore<K_Type extends ItemType, T_ReturnedRepondProps>(
       run: rerender,
       runAtStart: false, // runAtStart false since it's returning the initial state already, no need to set state
       ...options,
-    });
+    } as any);
 
     return () => stopEffect(effectId);
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -32,8 +32,8 @@ export function useStore<K_Type extends ItemType, T_ReturnedRepondProps>(
 }
 
 export function useStoreEffect<K_Type extends ItemType>(
-  run: Effect["run"],
-  options: Omit<Effect, "run">,
+  run: EffectDef["run"],
+  options: Omit<EffectDef, "run">,
   hookDeps: any[] | undefined = undefined
 ) {
   // const stringifiedCheck = JSON.stringify(check); // NOTE this may be bad for memory and performance, and might be better for people to have to manually update deps
@@ -41,7 +41,7 @@ export function useStoreEffect<K_Type extends ItemType>(
   useLayoutEffect(
     () => {
       const effectId = toSafeEffectId("useStoreEffect_" + JSON.stringify(options.changes));
-      startNewEffect({ id: effectId, atStepEnd: true, run, runAtStart: true, ...options }); // runAtStart true so it works like useEffect
+      startNewEffect({ id: effectId, atStepEnd: true, run, runAtStart: true, ...options } as any); // runAtStart true so it works like useEffect
       return () => stopEffect(effectId);
     },
     hookDeps ? [...hookDeps, stringifiedCheck] : [stringifiedCheck]